flask_app/requirements.txt
Flask
python-dotenv
paho-mqtt
flask_cors
Flask
request
flask_socketio
flask_cors